:sparkles: F06 > ToggleButton > create Toast in reusable-ui folder
